+++ src/closure-externs/closure-externs.js
@@ -999,7 +999,6 @@ var noExitRuntime;
 //    https://github.com/google/closure-compiler/issues/3167
 var BigInt;
 
-
 // Worklet
 /**
  * @constructor